[#239]refactor: member entity 위치 변경
