+++ src/main/java/com/runninghi/AdminInitializer.java
@@ -1,6 +1,6 @@
 package com.runninghi;
 
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import com.runninghi.member.command.domain.aggregate.entity.enumtype.Role;
 import com.runninghi.member.command.domain.repository.MemberCommandRepository;
 import lombok.RequiredArgsConstructor;
