+++ src/main/java/com/runninghi/configuration/SecurityConfig.java
@@ -18,7 +18,8 @@ import org.springframework.security.web.authentication.www.BasicAuthenticationFi
 @Configuration
 public class SecurityConfig {
     private static final String[] ALLOWED_URIS = {"/api/v1/sign-up", "/api/v1/sign-in", "/api/v1/verify-duplication-id"
-            , "/api/v1/member/find-account", "/api/v1/member/find-password", "/api/v1/member/update-password"};
+            , "/api/v1/verify-duplication-nickname", "/api/v1/member/find-account", "/api/v1/member/find-password"
+            , "/api/v1/member/update-password"};
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
     private final AuthenticationEntryPoint entryPoint;
