+++ src/main/java/com/runninghi/member/command/application/dto/member/MemberInfoDTO.java
@@ -1,6 +1,6 @@
 package com.runninghi.member.command.application.dto.member;
 
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
