+++ src/main/java/com/runninghi/member/command/domain/repository/MemberCommandRepository.java
@@ -1,6 +1,6 @@
 package com.runninghi.member.command.domain.repository;
 
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
