+++ src/test/java/com/runninghi/feedback/command/service/FeedbackCommandServiceTests.java
@@ -12,7 +12,7 @@ import com.runninghi.feedback.command.domain.aggregate.entity.FeedbackCategory;
 import com.runninghi.feedback.command.domain.aggregate.vo.FeedbackWriterVO;
 import com.runninghi.feedback.command.domain.repository.FeedbackCommandRepository;
 import com.runninghi.feedback.command.infrastructure.service.ApiFeedbackCommandInfraService;
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import com.runninghi.member.command.domain.aggregate.entity.enumtype.Role;
 import com.runninghi.member.command.domain.repository.MemberCommandRepository;
 import jakarta.transaction.Transactional;
