+++ src/test/java/com/runninghi/keyword/command/application/service/KeywordCommandServiceTest.java
@@ -5,7 +5,7 @@ import com.runninghi.keyword.command.application.dto.request.KeywordUpdateReques
 import com.runninghi.keyword.command.application.dto.response.KeywordResponse;
 import com.runninghi.keyword.command.domain.aggregate.entity.Keyword;
 import com.runninghi.keyword.command.domain.repository.KeywordCommandRepository;
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import com.runninghi.member.command.domain.aggregate.entity.enumtype.Role;
 import com.runninghi.member.command.domain.repository.MemberCommandRepository;
 import org.assertj.core.api.Assertions;
