+++ src/test/java/com/runninghi/member/query/application/service/AdminQueryServiceTest.java
@@ -1,7 +1,7 @@
 package com.runninghi.member.query.application.service;
 
 
-import com.runninghi.member.command.domain.aggregate.Member;
+import com.runninghi.member.command.domain.aggregate.entity.Member;
 import com.runninghi.member.command.domain.aggregate.entity.enumtype.Role;
 import com.runninghi.member.command.domain.repository.MemberCommandRepository;
 import com.runninghi.member.query.application.dto.member.response.MemberInfoResponse;
